Final Commit (Error Handling, Store Completed)
1. Addressed errors when the user tries to update their account
2. Updating account now allows the user to modify their ID
3. Store integration can now properly withdraw money from the user's bank account
4. Store integration has a new feature where the user can clear their cart
5. Addressed various input validation related errors (for example, the user may be able to press 'enter' and skip validation, causing the program to crash)

+++ src/store_system/items/cpp/Product.cpp
@@ -19,7 +19,7 @@ std::string Product::getProductName() const {
 	return productName;
 }
 
-double Product::getPrice() const {
+float Product::getPrice() const {
 	return price;
 }
